+++ R/elo.glm.R
@@ -56,6 +56,8 @@ elo.glm <- function(formula, data, weights, na.action, subset, family = "binomia
   wts <- mf$weights
   dat.glm <- stats::glm(wins.A ~ ., data = dat, family = family, na.action = stats::na.pass, subset = NULL, weights = wts, ...)
   dat.glm$na.action <- stats::na.action(mf)
+  dat.glm$elo.terms <- Terms
+  dat.glm$teams <- all.teams
 
   if(running)
   {